backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.3
pydantic-settings==2.3.1
pypdf==4.2.0
python-docx==1.1.2
langchain==0.2.5
langchain-community==0.2.5
langchain-openai==0.1.9
langchain-anthropic==0.1.15
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
tiktoken==0.7.0
langchain-google-genai==1.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy>=2.0.0
